feat(ai): 更新聊天代理工厂以支持模型温度参数
- 在 ChatAgentFactory 构造函数中注入 IServiceProvider
- 修改模型初始化逻辑以传递 Temperature 参数
- 更新模板系统提示词属性引用
- 调整命名空间引用以支持新的 AI 扩展功能

+++ src/Zilor.AICopilot.AiGatewayService/Commands/ConversationTemplates/CreateConversationTemplate.cs
@@ -18,7 +18,7 @@ public record CreateConversationTemplateCommand(
     string SystemPrompt,
     Guid ModelId,
     int? MaxTokens,
-    double? Temperature) : ICommand<Result<CreatedConversationTemplateDto>>;
+    float? Temperature) : ICommand<Result<CreatedConversationTemplateDto>>;
 
 public class CreateConversationTemplateCommandHandler(IRepository<ConversationTemplate> repo)
     : ICommandHandler<CreateConversationTemplateCommand, Result<CreatedConversationTemplateDto>>
