+++ src/Zilor.AICopilot.AiGatewayService/Commands/LanguageModels/CreateLanguageModel.cs
@@ -18,7 +18,7 @@ public record CreateLanguageModelCommand(
     string BaseUrl,
     string? ApiKey,
     int MaxTokens,
-    double Temperature = 0.7) : ICommand<Result<CreatedLanguageModelDto>>;
+    float Temperature = 0.7f) : ICommand<Result<CreatedLanguageModelDto>>;
 
 public class CreateLanguageModelCommandHandler(IRepository<LanguageModel> repo)
     : ICommandHandler<CreateLanguageModelCommand, Result<CreatedLanguageModelDto>>
